Render spot colors in ImageStream
