+++ crates/jxl-oxide-cli/src/decode.rs
@@ -99,7 +99,7 @@ pub fn handle_decode(args: DecodeArgs) -> Result<()> {
     let mps = total_pixels as f64 / 1e6;
 
     if args.output_format == OutputFormat::Npy {
-        image.set_render_spot_colour(false);
+        image.set_render_spot_color(false);
     }
 
     let keyframes = if let Some(num_reps @ 2..) = args.num_reps {
