+++ crates/jxl-oxide/src/lib.rs
@@ -172,7 +172,7 @@ mod lcms2;
 
 #[cfg(feature = "lcms2")]
 pub use self::lcms2::Lcms2;
-pub use fb::FrameBuffer;
+pub use fb::{FrameBuffer, ImageStream};
 
 pub type Result<T> = std::result::Result<T, Box<dyn std::error::Error + Send + Sync + 'static>>;
 
@@ -386,7 +386,7 @@ impl UninitializedJxlImage {
         let bytes_read = bitstream.num_read_bits() / 8 + skip_bytes;
         self.buffer.drain(..bytes_read);
 
-        let render_spot_colour = !image_header.metadata.grayscale();
+        let render_spot_color = !image_header.metadata.grayscale();
 
         let mut builder = RenderContext::builder().pool(self.pool.clone());
         if let Some(icc) = embedded_icc {
@@ -405,7 +405,7 @@ impl UninitializedJxlImage {
             reader: self.reader,
             image_header,
             ctx,
-            render_spot_colour,
+            render_spot_color,
             end_of_image: false,
             buffer: Vec::new(),
             buffer_offset: bytes_read,
@@ -433,7 +433,7 @@ pub struct JxlImage {
     reader: ContainerDetectingReader,
     image_header: Arc<ImageHeader>,
     ctx: RenderContext,
-    render_spot_colour: bool,
+    render_spot_color: bool,
     end_of_image: bool,
     buffer: Vec<u8>,
     buffer_offset: usize,
@@ -641,18 +641,18 @@ impl JxlImage {
 
     /// Returns whether the spot color channels will be rendered.
     #[inline]
-    pub fn render_spot_colour(&self) -> bool {
-        self.render_spot_colour
+    pub fn render_spot_color(&self) -> bool {
+        self.render_spot_color
     }
 
     /// Sets whether the spot colour channels will be rendered.
     #[inline]
-    pub fn set_render_spot_colour(&mut self, render_spot_colour: bool) -> &mut Self {
-        if render_spot_colour && self.image_header.metadata.grayscale() {
+    pub fn set_render_spot_color(&mut self, render_spot_color: bool) -> &mut Self {
+        if render_spot_color && self.image_header.metadata.grayscale() {
             tracing::warn!("Spot colour channels are not rendered on grayscale images");
             return self;
         }
-        self.render_spot_colour = render_spot_colour;
+        self.render_spot_color = render_spot_color;
         self
     }
 
@@ -739,6 +739,7 @@ impl JxlImage {
             extra_channels: self.convert_ec_info(),
             target_frame_region,
             color_bit_depth: self.image_header.metadata.bit_depth,
+            render_spot_color: self.render_spot_color,
         };
         Ok(result)
     }
@@ -772,6 +773,7 @@ impl JxlImage {
             extra_channels: self.convert_ec_info(),
             target_frame_region,
             color_bit_depth: self.image_header.metadata.bit_depth,
+            render_spot_color: self.render_spot_color,
         };
         Ok(result)
     }
@@ -889,6 +891,7 @@ pub struct Render {
     extra_channels: Vec<ExtraChannel>,
     target_frame_region: Region,
     color_bit_depth: BitDepth,
+    render_spot_color: bool,
 }
 
 impl Render {
@@ -1050,166 +1053,7 @@ impl Render {
     /// The stream will include black and alpha channels, if exists, in addition to color channels.
     /// Orientation is applied.
     pub fn stream(&self) -> ImageStream {
-        let orientation = self.orientation;
-        assert!((1..=8).contains(&orientation));
-        let Region {
-            left,
-            top,
-            mut width,
-            mut height,
-        } = self.target_frame_region;
-        if orientation >= 5 {
-            std::mem::swap(&mut width, &mut height);
-        }
-
-        let fb = self.image.buffer();
-        let color_channels = self.image.color_channels();
-        let regions_and_shifts = self.image.regions_and_shifts();
-
-        let mut grids: Vec<_> = self.color_channels().iter().collect();
-        let mut bit_depth = vec![self.color_bit_depth; grids.len()];
-
-        let mut start_offset_xy = Vec::new();
-        for (region, _) in &regions_and_shifts[..color_channels] {
-            start_offset_xy.push((left - region.left, top - region.top));
-        }
-
-        // Find black
-        for (ec_idx, (ec, (region, _))) in self
-            .extra_channels
-            .iter()
-            .zip(&regions_and_shifts[color_channels..])
-            .enumerate()
-        {
-            if ec.is_black() {
-                grids.push(&fb[color_channels + ec_idx]);
-                bit_depth.push(ec.bit_depth);
-                start_offset_xy.push((left - region.left, top - region.top));
-                break;
-            }
-        }
-        // Find alpha
-        for (ec_idx, (ec, (region, _))) in self
-            .extra_channels
-            .iter()
-            .zip(&regions_and_shifts[color_channels..])
-            .enumerate()
-        {
-            if ec.is_alpha() {
-                grids.push(&fb[color_channels + ec_idx]);
-                bit_depth.push(ec.bit_depth);
-                start_offset_xy.push((left - region.left, top - region.top));
-                break;
-            }
-        }
-
-        ImageStream {
-            orientation,
-            width,
-            height,
-            grids,
-            bit_depth,
-            start_offset_xy,
-            y: 0,
-            x: 0,
-            c: 0,
-        }
-    }
-}
-
-/// Image stream that writes to borrowed buffer.
-pub struct ImageStream<'r> {
-    orientation: u32,
-    width: u32,
-    height: u32,
-    grids: Vec<&'r ImageBuffer>,
-    start_offset_xy: Vec<(i32, i32)>,
-    bit_depth: Vec<BitDepth>,
-    y: u32,
-    x: u32,
-    c: u32,
-}
-
-impl ImageStream<'_> {
-    /// Returns width of the image.
-    #[inline]
-    pub fn width(&self) -> u32 {
-        self.width
-    }
-
-    /// Returns height of the image.
-    #[inline]
-    pub fn height(&self) -> u32 {
-        self.height
-    }
-
-    /// Returns the number of channels of the image.
-    #[inline]
-    pub fn channels(&self) -> u32 {
-        self.grids.len() as u32
-    }
-
-    /// Writes next samples to the buffer, returning how many samples are written.
-    pub fn write_to_buffer(&mut self, buf: &mut [f32]) -> usize {
-        let channels = self.grids.len() as u32;
-        let mut buf_it = buf.iter_mut();
-        let mut count = 0usize;
-        'outer: while self.y < self.height {
-            while self.x < self.width {
-                while self.c < channels {
-                    let Some(v) = buf_it.next() else {
-                        break 'outer;
-                    };
-                    let (start_x, start_y) = self.start_offset_xy[self.c as usize];
-                    let (x, y) = self.to_original_coord(self.x, self.y);
-                    let (Some(x), Some(y)) =
-                        (x.checked_add_signed(start_x), y.checked_add_signed(start_y))
-                    else {
-                        *v = 0.0;
-                        count += 1;
-                        self.c += 1;
-                        continue;
-                    };
-                    let x = x as usize;
-                    let y = y as usize;
-                    let grid = &self.grids[self.c as usize];
-                    let bit_depth = self.bit_depth[self.c as usize];
-                    *v = match grid {
-                        ImageBuffer::F32(g) => g.get(x, y).copied().unwrap_or(0.0),
-                        ImageBuffer::I32(g) => {
-                            bit_depth.parse_integer_sample(g.get(x, y).copied().unwrap_or(0))
-                        }
-                        ImageBuffer::I16(g) => {
-                            bit_depth.parse_integer_sample(g.get(x, y).copied().unwrap_or(0) as i32)
-                        }
-                    };
-                    count += 1;
-                    self.c += 1;
-                }
-                self.c = 0;
-                self.x += 1;
-            }
-            self.x = 0;
-            self.y += 1;
-        }
-        count
-    }
-
-    #[inline]
-    fn to_original_coord(&self, x: u32, y: u32) -> (u32, u32) {
-        let width = self.width;
-        let height = self.height;
-        match self.orientation {
-            1 => (x, y),
-            2 => (width - x - 1, y),
-            3 => (width - x - 1, height - y - 1),
-            4 => (x, height - y - 1),
-            5 => (y, x),
-            6 => (y, width - x - 1),
-            7 => (height - y - 1, width - x - 1),
-            8 => (height - y - 1, x),
-            _ => unreachable!(),
-        }
+        ImageStream::from_render(self)
     }
 }
